Factor LOD into a centralized manager
