+++ src/viewer/Viewer.js
@@ -52,6 +52,8 @@ class Viewer {
      * that it does not clip huge models.
      * @param {Boolean} [cfg.colorTextureEnabled=true] Whether to enable base color texture rendering.
      * @param {Boolean} [cfg.pbrEnabled=false] Whether to enable physically-based rendering.
+     * @param {Boolean} [cfg.lodEnabled=false] Whether to enable Level-of-Detail (LOD) culling. See {@link LOD} for more info.
+     * @param {Boolean} [cfg.vfcEnabled=false] Whether to enable View Frustum Culling (VFC) culling. See {@link VFC} for more info.
      * @param {LocaleService} [cfg.localeService=null] Optional locale-based translation service.
      */
     constructor(cfg) {
@@ -108,6 +110,8 @@ class Viewer {
             entityOffsetsEnabled: (!!cfg.entityOffsetsEnabled),
             logarithmicDepthBufferEnabled: (!!cfg.logarithmicDepthBufferEnabled),
             pbrEnabled: (!!cfg.pbrEnabled),
+            lodEnabled: (!!cfg.lodEnabled),
+            vfcCulling: (!!cfg.vfcEnabled),
             colorTextureEnabled: (cfg.colorTextureEnabled !== false)
         });
